# swagger/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is a demo API"
  version: "1.0.0"
  title: "Simple Burger Menu API"
  contact:
    email: "james.reinknecht@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "jamesreinknecht.com"
basePath: "/burgers"
...
schemes:
- "https"
paths:
...
  /eatburger/{id}:
    put:
      summary: "Updates burgers to is_eaten to true in the database"
      description: "Page reloads automagically with jquery from html page,"
      operationId: "eatburgerIdPUT"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Numeric ID of the burger to update."
        required: true
        type: "integer"
      responses:
        200:
          description: "OK"
      x-swagger-router-controller: "Default"
...
  /delete/{id}:
    delete:
      summary: "deletes a particular burgerid"
      operationId: "deleteIdDELETE"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Numeric ID of the burger to get."
        required: true
        type: "integer"
      responses:
        200:
          description: "OK"
      x-swagger-router-controller: "Default"
...